Refactor overload (#2432)
* Abstracting plugin overload logic

* Abstracting plugin overload logic

* add use strict
